+++ CHANGELOG.md
@@ -8,4 +8,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - `omnipytent.ext.extra.screen` module for interaction with [GNU Screen](https://www.gnu.org/software/screen/)
-- `omnipytent.ext.extra.testing` module for test choosing helpers.
+- `omnipytent.ext.extra.testing` subpackage for test choosing helpers.
+- `omnipytent.ext.extra.repls` subpackage for interacting with REPLs. Currently has one submodule for `ipython`